Search bar and when you click each hot sauce, something pops up
