+++ group-project-ui/src/modules/pages/pages.config.js
@@ -9,10 +9,10 @@ const TestComp = () => (
 // NOTE: Including linkText property will include route in navigation menu; omit this property for routes which shouldn't be in navigation
 // NOTE: You may need to update this config with more properties and map to the RootLayout if more complex routes are required
 export const routes = [
-  { path: '/test', linkText: 'Test', routeComponent: TestComp, exact: true },
+  //{ path: '/test', linkText: 'Test', routeComponent: TestComp, exact: true },
   // Currently a fallback route which will match if any routes above are not exactly matched
   // NOTE: Fallback routes should always be last
-  { path: '/', linkText: 'Home', routeComponent: HomePageComponent, exact: false },
+  { path: '/', linkText: 'Hot Sauce Main Directory and Search', routeComponent: HomePageComponent, exact: false },
 ];
 
 export default routes;
